Refactored to_singular to be cleaner
Used method extraction to reduce the number of lines per method and
added tests around each of the private methods to ensure that they
function as expected. Test coverage is not 100% on the code but should
provide a baseline for future refactors.

The trade off that was taken is that a couple of the methods now return
an Option<String> since the code cannot assert that it will always
return a String. This means that patching this part of the code will
require understanding of the Option type in Rust.