Moved heavy code to a feature flag that will permit turning those features off.
Some projects have requested a feature flag system for Inflector that
would permit turning off some features. This is an issue for example
with pluralize and singularize which are generally not used but add a
high cost to compile times. The solution proposed fixes that issue by
placing a large portion of the heavy code under a feature flag, and
making the dependencies optional for those not using the heavier
methods.

The side effect in this case is added complexity for some newer users,
maybe.